+++ packages/polar-betterauth/example/src/lib/auth-client.ts
@@ -4,7 +4,7 @@ import { organizationClient } from "better-auth/client/plugins";
 
 export const authClient = createAuthClient({
 	baseURL: "http://localhost:3001", // the base url of your auth server
-	plugins: [organizationClient(), polarClient({ checkout: true })],
+	plugins: [organizationClient(), polarClient()],
 });
 
 export const { signIn, signUp, useSession } = authClient;
