+++ packages/polar-betterauth/package.json
@@ -32,7 +32,7 @@
 		"@polar-sh/sdk": "^0.32.13",
 		"@sindresorhus/tsconfig": "^7.0.0",
 		"@types/node": "^20.0.0",
-		"better-auth": "^1.2.0",
+		"better-auth": "^1.2.7",
 		"tsup": "^8.3.5",
 		"vitest": "^2.1.8"
 	},
@@ -41,6 +41,6 @@
 	},
 	"peerDependencies": {
 		"@polar-sh/sdk": "^0.32.13",
-		"better-auth": "^1.2.0"
+		"better-auth": "^1.2.7"
 	}
 }
