+++ packages/polar-betterauth/src/client.ts
@@ -1,29 +1,9 @@
 import type { BetterAuthClientPlugin } from "better-auth";
 import type { polar } from "./index";
 
-export const polarClient = <
-	O extends {
-		checkout: boolean;
-	},
->(
-	_options?: O,
-) => {
+export const polarClient = () => {
 	return {
 		id: "polar-client",
-		$InferServerPlugin: {} as ReturnType<
-			typeof polar<
-				O["checkout"] extends true
-					? {
-							client: any;
-							checkout: {
-								enabled: true;
-								products: [];
-							};
-						}
-					: {
-							client: any;
-						}
-			>
-		>,
+		$InferServerPlugin: {} as ReturnType<typeof polar>,
 	} satisfies BetterAuthClientPlugin;
 };
