+++ packages/polar-betterauth/src/plugins/checkout.ts
@@ -1,0 +1,105 @@
+import { APIError, getSessionFromCtx } from "better-auth/api";
+import { createAuthEndpoint } from "better-auth/plugins";
+import { z } from "zod";
+import type { Polar } from "@polar-sh/sdk";
+import type { Product } from "../types";
+
+export interface CheckoutOptions {
+	/**
+	 * Optional list of slug -> productId mappings for easy slug checkouts
+	 */
+	products?: Product[] | (() => Promise<Product[]>);
+	/**
+	 * Checkout Success URL
+	 */
+	successUrl?: string;
+	/**
+	 * Only allow authenticated customers to checkout
+	 */
+	authenticatedUsersOnly?: boolean;
+}
+
+export const checkout =
+	(checkoutOptions: CheckoutOptions = {}) =>
+	(polar: Polar) => {
+		return {
+			checkout: createAuthEndpoint(
+				"/checkout",
+				{
+					method: "POST",
+					body: z.object({
+						products: z.union([z.array(z.string()), z.string()]).optional(),
+						slug: z.string().optional(),
+						referenceId: z.string().optional(),
+					}),
+				},
+				async (ctx) => {
+					const session = await getSessionFromCtx(ctx);
+
+					let productIds: string[] = [];
+
+					if (ctx.body.slug) {
+						const resolvedProducts = await (typeof checkoutOptions.products ===
+						"function"
+							? checkoutOptions.products()
+							: checkoutOptions.products);
+
+						const productId = resolvedProducts?.find(
+							(product) => product.slug === ctx.body.slug,
+						)?.productId;
+
+						if (!productId) {
+							throw new APIError("BAD_REQUEST", {
+								message: "Product not found",
+							});
+						}
+
+						productIds = [productId];
+					} else {
+						productIds = Array.isArray(ctx.body.products)
+							? ctx.body.products.filter((id) => id !== undefined)
+							: [ctx.body.products].filter((id) => id !== undefined);
+					}
+
+					if (checkoutOptions.authenticatedUsersOnly && !session?.user.id) {
+						throw new APIError("UNAUTHORIZED", {
+							message: "You must be logged in to checkout",
+						});
+					}
+
+					try {
+						const checkout = await polar.checkouts.create({
+							customerExternalId: session?.user.id,
+							products: productIds,
+							successUrl: checkoutOptions.successUrl
+								? new URL(
+										checkoutOptions.successUrl,
+										ctx.request?.url,
+									).toString()
+								: undefined,
+							metadata: ctx.body.referenceId
+								? {
+										referenceId: ctx.body.referenceId,
+									}
+								: undefined,
+						});
+
+						return ctx.json({
+							url: checkout.url,
+							redirect: true,
+						});
+					} catch (e: unknown) {
+						if (e instanceof Error) {
+							ctx.context.logger.error(
+								`Polar checkout creation failed. Error: ${e.message}`,
+							);
+						}
+
+						throw new APIError("INTERNAL_SERVER_ERROR", {
+							message: "Checkout creation failed",
+						});
+					}
+				},
+			),
+		};
+	};
