+++ packages/polar-betterauth/src/plugins/portal.ts
@@ -1,0 +1,171 @@
+import type { Polar } from "@polar-sh/sdk";
+import { APIError } from "better-auth/api";
+import { sessionMiddleware } from "better-auth/api";
+import { createAuthEndpoint } from "better-auth/plugins";
+import { z } from "zod";
+
+export const portal = () => (polar: Polar) => {
+	return {
+		portal: createAuthEndpoint(
+			"/customer/portal",
+			{
+				method: "GET",
+				use: [sessionMiddleware],
+			},
+			async (ctx) => {
+				if (!ctx.context.session?.user.id) {
+					throw new APIError("BAD_REQUEST", {
+						message: "User not found",
+					});
+				}
+
+				try {
+					const customerSession = await polar.customerSessions.create({
+						customerExternalId: ctx.context.session?.user.id,
+					});
+
+					return ctx.json({
+						url: customerSession.customerPortalUrl,
+						redirect: true,
+					});
+				} catch (e: unknown) {
+					if (e instanceof Error) {
+						ctx.context.logger.error(
+							`Polar customer portal creation failed. Error: ${e.message}`,
+						);
+					}
+
+					throw new APIError("INTERNAL_SERVER_ERROR", {
+						message: "Customer portal creation failed",
+					});
+				}
+			},
+		),
+		state: createAuthEndpoint(
+			"/customer/state",
+			{
+				method: "GET",
+				use: [sessionMiddleware],
+			},
+			async (ctx) => {
+				if (!ctx.context.session.user.id) {
+					throw new APIError("BAD_REQUEST", {
+						message: "User not found",
+					});
+				}
+
+				try {
+					const state = await polar.customers.getStateExternal({
+						externalId: ctx.context.session?.user.id,
+					});
+
+					return ctx.json(state);
+				} catch (e: unknown) {
+					if (e instanceof Error) {
+						ctx.context.logger.error(
+							`Polar subscriptions list failed. Error: ${e.message}`,
+						);
+					}
+
+					throw new APIError("INTERNAL_SERVER_ERROR", {
+						message: "Subscriptions list failed",
+					});
+				}
+			},
+		),
+		benefits: createAuthEndpoint(
+			"/customer/benefits/list",
+			{
+				method: "GET",
+				query: z
+					.object({
+						page: z.number().optional(),
+						limit: z.number().optional(),
+					})
+					.optional(),
+				use: [sessionMiddleware],
+			},
+			async (ctx) => {
+				if (!ctx.context.session.user.id) {
+					throw new APIError("BAD_REQUEST", {
+						message: "User not found",
+					});
+				}
+
+				try {
+					const customerSession = await polar.customerSessions.create({
+						customerExternalId: ctx.context.session?.user.id,
+					});
+
+					const benefits = await polar.customerPortal.benefitGrants.list(
+						{ customerSession: customerSession.token },
+						{
+							page: ctx.query?.page,
+							limit: ctx.query?.limit,
+						},
+					);
+
+					return ctx.json(benefits);
+				} catch (e: unknown) {
+					if (e instanceof Error) {
+						ctx.context.logger.error(
+							`Polar benefits list failed. Error: ${e.message}`,
+						);
+					}
+
+					throw new APIError("INTERNAL_SERVER_ERROR", {
+						message: "Benefits list failed",
+					});
+				}
+			},
+		),
+		orders: createAuthEndpoint(
+			"/customer/orders/list",
+			{
+				method: "GET",
+				query: z
+					.object({
+						page: z.number().optional(),
+						limit: z.number().optional(),
+						productBillingType: z.enum(["recurring", "one_time"]).optional(),
+					})
+					.optional(),
+				use: [sessionMiddleware],
+			},
+			async (ctx) => {
+				if (!ctx.context.session.user.id) {
+					throw new APIError("BAD_REQUEST", {
+						message: "User not found",
+					});
+				}
+
+				try {
+					const customerSession = await polar.customerSessions.create({
+						customerExternalId: ctx.context.session?.user.id,
+					});
+
+					const orders = await polar.customerPortal.orders.list(
+						{ customerSession: customerSession.token },
+						{
+							page: ctx.query?.page,
+							limit: ctx.query?.limit,
+							productBillingType: ctx.query?.productBillingType,
+						},
+					);
+
+					return ctx.json(orders);
+				} catch (e: unknown) {
+					if (e instanceof Error) {
+						ctx.context.logger.error(
+							`Polar orders list failed. Error: ${e.message}`,
+						);
+					}
+
+					throw new APIError("INTERNAL_SERVER_ERROR", {
+						message: "Orders list failed",
+					});
+				}
+			},
+		),
+	};
+};
