+++ packages/polar-betterauth/src/plugins/usage.ts
@@ -1,0 +1,106 @@
+import type { Polar } from "@polar-sh/sdk";
+import type { Product } from "../types";
+import {
+	APIError,
+	createAuthEndpoint,
+	sessionMiddleware,
+} from "better-auth/api";
+import { z } from "zod";
+
+export interface UsageOptions {
+	/**
+	 * Products to use for topping up credits
+	 */
+	creditProducts?: Product[] | (() => Promise<Product[]>);
+}
+
+export const usage = (_usageOptions?: UsageOptions) => (polar: Polar) => {
+	return {
+		meters: createAuthEndpoint(
+			"/usage/meters/list",
+			{
+				method: "GET",
+				use: [sessionMiddleware],
+				query: z.object({
+					page: z.number().optional(),
+					limit: z.number().optional(),
+				}),
+			},
+			async (ctx) => {
+				if (!ctx.context.session.user.id) {
+					throw new APIError("BAD_REQUEST", {
+						message: "User not found",
+					});
+				}
+
+				try {
+					const customerSession = await polar.customerSessions.create({
+						customerExternalId: ctx.context.session.user.id,
+					});
+
+					const customerMeters = await polar.customerPortal.customerMeters.list(
+						{ customerSession: customerSession.token },
+						{
+							page: ctx.query?.page,
+							limit: ctx.query?.limit,
+						},
+					);
+
+					return ctx.json(customerMeters);
+				} catch (e: unknown) {
+					if (e instanceof Error) {
+						ctx.context.logger.error(
+							`Polar meters list failed. Error: ${e.message}`,
+						);
+					}
+
+					throw new APIError("INTERNAL_SERVER_ERROR", {
+						message: "Meters list failed",
+					});
+				}
+			},
+		),
+		ingestion: createAuthEndpoint(
+			"/usage/ingest",
+			{
+				method: "POST",
+				body: z.object({
+					event: z.string(),
+					metadata: z.record(z.any()),
+				}),
+				use: [sessionMiddleware],
+			},
+			async (ctx) => {
+				if (!ctx.context.session.user.id) {
+					throw new APIError("BAD_REQUEST", {
+						message: "User not found",
+					});
+				}
+
+				try {
+					const ingestion = await polar.events.ingest({
+						events: [
+							{
+								name: ctx.body.event,
+								metadata: ctx.body.metadata,
+								externalCustomerId: ctx.context.session.user.id,
+							},
+						],
+					});
+
+					return ctx.json(ingestion);
+				} catch (e: unknown) {
+					if (e instanceof Error) {
+						ctx.context.logger.error(
+							`Polar ingestion failed. Error: ${e.message}`,
+						);
+					}
+
+					throw new APIError("INTERNAL_SERVER_ERROR", {
+						message: "Ingestion failed",
+					});
+				}
+			},
+		),
+	};
+};
